moved microscope types into from_lims
moved microscope type dict and get_microscope_type from the from_lims_utilities module into from_lims
